Simplify deserialization of Item
As a reading item could have a resolved url that was extracted from a
Tweet, for example, we have to pick the resolved when when available.

This was happening on the deserialization method, which was being manual
written.
Instead, we now let the derived implementation extract the information
when available and use a method to either select that or the given url.