Accepts a more generic structure on the interface
Slices are a good abstraction to take interfaces, specially when you
need indexing, but it is even better to take IntoIterator when you just
need to transform information.

This will let us take instead of just Vec's converted into slices, Sets
or other structures that now how to be iterated linearly.