Cleanup URLs from hashes
Some websites uses the hash or query strings to track where the user
came from, which campaign provided the link and so on.

Some examples:

- Medium
- Buffer twitter posts
- Blogspot (which changes the domain completely)

This commit try to address this by keeping another set with cleaned up
URls, which has no query strings or hash.

The idea to use a set is to exchange memory for speed, as it will locate
the content faster than looping all the time. This will only be used if
the as-given lookup fails.