Uses multiple PHONY targets to better structure the file
Instead of grouping all the dependencies on a single .PHONY target, it
is possible to declar multiple targets on the file, and `make` will
collect them all.

This way, we can easily see that a target is phony close to the
declaration, and possibly use it to automatically generate the help
documentation.