Decouple the method interface from the response format
The response from the API does not return the full list, as I've reached
the upper limit per response (5000). The method to return all should be
deoupled from the response format, as now it is needed to inspect the
search_metadata object, and this would be exposing too much from the API
response.

This commit replaces the return of response to an public type alias.
This will allow us to extend the content until everything is done.