Cleanup https redirection
Sometimes, the link sent to Pocket is stored as http, but on access by
the browser, we are redirected to an https version.

The cleanup will now fix the scheme, instead of using what was stored.
This way, when trying to mark a url as read, we can match the schemas.

An alternative would be to strip the schema completely, so we don't
direct users to the wrong website on clean. Since I'm only using the
cleanup function to match urls with ids, this is not much of a problem.